+++ src/main/kotlin/uk/co/jcox/chemvis/application/moleditor/actions/EditorAction.kt
@@ -62,6 +62,35 @@ abstract class EditorAction {
         fakeLabel.addComponent(GhostImplicitHydrogenGroupComponent())
     }
 
+
+    protected fun updateGhostGroups(molManager: IMoleculeManager, levelAtom: EntityLevel, structAtom: UUID) {
+
+        removeImplicitHydrogenGroup(levelAtom)
+
+        if (!levelAtom.hasComponent(AlwaysExplicit::class) && molManager.isOfElement(structAtom, "C")) {
+            return
+        }
+
+        val newHydrogenCount = molManager.getImplicitHydrogens(structAtom)
+
+        insertImplicitHydrogenGroup(levelAtom, newHydrogenCount)
+    }
+
+
+    protected fun makeCarbonImplicit(molManager: IMoleculeManager, structMol: UUID, structCarbon: UUID, levelCarbon: EntityLevel) {
+        //First check if carbon
+        if (! molManager.isOfElement(structCarbon, "C")) {
+            return
+        }
+
+        val bonds = molManager.getBonds(structMol, structCarbon)
+
+        if (bonds >= NewOrganicEditorState.CARBON_IMPLICIT_LIMIT) {
+            //Then hide the text component of the carbon
+            val textComp = levelCarbon.getComponent(TransformComponent::class)
+            textComp.visible = false
+        }
+    }
 }
